# docs/swagger.yaml
...
info:
  contact:
    email: self@gimo.me
    name: Yuanji
    url: https://gimo.me
  description: A simple HTTP Request & Response Service.
  license:
    name: MIT License
    url: https://github.com/masakichi/echobin/blob/main/LICENSE
  title: echobin API
  version: "0.1"
paths:
  /absolute-redirect/{n}:
    get:
      parameters:
      - description: "n"
        in: path
        name: "n"
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "302":
          description: A redirection.
      summary: Absolutely 302 Redirects n times.
      tags:
      - Redirects
...
  /base64/{value}:
    get:
      parameters:
      - default: RUNIT0JJTiBpcyBhd2Vzb21l
        description: Encoded base64 content
        in: path
        name: value
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Decoded base64 content.
      summary: Decodes base64url-encoded string.
      tags:
      - Dynamic data
...
  /bytes/{n}:
    get:
      parameters:
      - description: number of bytes
        in: path
        name: "n"
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Bytes.
      summary: Returns n random bytes generated with given seed
      tags:
      - Dynamic data
...
  /cookies/delete:
    get:
      parameters:
      - description: freeform
        in: query
        name: freeform
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Redirect to cookie list
      summary: Deletes cookie(s) as provided by the query string and redirects to
        cookie list.
      tags:
      - Cookies
  /cookies/set:
    get:
      parameters:
      - description: freeform
        in: query
        name: freeform
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Redirect to cookie list
      summary: Sets cookie(s) as provided by the query string and redirects to cookie
        list.
      tags:
      - Cookies
  /cookies/set/{name}/{value}:
    get:
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      - description: value
        in: path
        name: value
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Set cookies and redirects to cookie list.
      summary: Sets a cookie and redirects to cookie list.
      tags:
      - Cookies
...
  /delay/{delay}:
    delete:
      parameters:
      - description: delay
        in: path
        name: delay
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: A delayed response.
      summary: Returns a delayed response (max of 10 seconds).
      tags:
      - Dynamic data
    get:
      parameters:
      - description: delay
        in: path
        name: delay
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: A delayed response.
      summary: Returns a delayed response (max of 10 seconds).
      tags:
      - Dynamic data
    patch:
      parameters:
      - description: delay
        in: path
        name: delay
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: A delayed response.
      summary: Returns a delayed response (max of 10 seconds).
      tags:
      - Dynamic data
    post:
      parameters:
      - description: delay
        in: path
        name: delay
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: A delayed response.
      summary: Returns a delayed response (max of 10 seconds).
      tags:
      - Dynamic data
    put:
      parameters:
      - description: delay
        in: path
        name: delay
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: A delayed response.
      summary: Returns a delayed response (max of 10 seconds).
      tags:
      - Dynamic data
...
  /drip:
    get:
      parameters:
      - default: 200
        description: The response code that will be returned
        in: query
        name: code
        type: integer
      - default: 2
        description: The amount of time (in seconds) to delay before responding
        in: query
        name: delay
        type: number
      - default: 2
        description: The amount of time (in seconds) over which to drip each byte
        in: query
        name: duration
        type: number
      - default: 10
        description: The number of bytes to respond with
        in: query
        name: numbytes
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: A dripped response.
      summary: Drips data over a duration after an optional initial delay.
      tags:
      - Dynamic data
...
  /links/{n}/{offset}:
    get:
      parameters:
      - default: 1
        description: The amount of links
        in: path
        name: "n"
        required: true
        type: integer
      - default: 0
        description: Offset starts from 0
        in: path
        name: offset
        required: true
        type: integer
      produces:
      - text/html
      responses:
        "200":
          description: HTML links.
      summary: Generate a page containing n links to other pages which do the same.
      tags:
      - Dynamic data
...
  /range/{numbytes}:
    get:
      deprecated: true
      parameters:
      - default: 10
        description: The amount of bytes
        in: path
        name: numbytes
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Bytes
      summary: Streams n random bytes generated with given seed, at given chunk size
        per packet.
      tags:
      - Dynamic data
  /redirect-to:
    delete:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: url
        in: formData
        name: url
        required: true
        type: string
      - description: status_code
        in: formData
        name: status_code
        type: integer
      produces:
      - text/plain
      responses:
        "302":
          description: A redirection.
      summary: 302/3XX Redirects to the given URL.
      tags:
      - Redirects
    get:
      parameters:
      - description: url
        in: query
        name: url
        required: true
        type: string
      - description: status_code
        in: query
        name: status_code
        type: integer
      produces:
      - text/plain
      responses:
        "302":
          description: A redirection.
      summary: 302/3XX Redirects to the given URL.
      tags:
      - Redirects
    patch:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: url
        in: formData
        name: url
        required: true
        type: string
      - description: status_code
        in: formData
        name: status_code
        type: integer
      produces:
      - text/plain
      responses:
        "302":
          description: A redirection.
      summary: 302/3XX Redirects to the given URL.
      tags:
      - Redirects
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: url
        in: formData
        name: url
        required: true
        type: string
      - description: status_code
        in: formData
        name: status_code
        type: integer
      produces:
      - text/plain
      responses:
        "302":
          description: A redirection.
      summary: 302/3XX Redirects to the given URL.
      tags:
      - Redirects
    put:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: url
        in: formData
        name: url
        required: true
        type: string
      - description: status_code
        in: formData
        name: status_code
        type: integer
      produces:
      - text/plain
      responses:
        "302":
          description: A redirection.
      summary: 302/3XX Redirects to the given URL.
      tags:
      - Redirects
  /redirect/{n}:
    get:
      parameters:
      - description: "n"
        in: path
        name: "n"
        required: true
        type: integer
      - default: "false"
        description: absolute
        in: query
        name: absolute
        type: string
      produces:
      - text/plain
      responses:
        "302":
          description: A redirection.
      summary: 302 Redirects n times.
      tags:
      - Redirects
  /relative-redirect/{n}:
    get:
      parameters:
      - description: "n"
        in: path
        name: "n"
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "302":
          description: A redirection.
      summary: Relatively 302 Redirects n times.
      tags:
      - Redirects
...
  /stream-bytes/{n}:
    get:
      deprecated: true
      parameters:
      - default: 10
        description: The amount of bytes
        in: path
        name: "n"
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Bytes
      summary: Streams n random bytes generated with given seed, at given chunk size
        per packet.
      tags:
      - Dynamic data
  /stream/{n}:
    get:
      parameters:
      - default: 10
        description: The amount of JSON objects
        in: path
        name: "n"
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Streamed JSON responses.
      summary: Stream n JSON responses
      tags:
      - Dynamic data
...
swagger: "2.0"
...